Refactor ZMQ streams into separate modules
